requirements.txt
discord.py==1.5.0
python-dotenv==0.17.1
requests==2.26.0
pillow==8.3.1